requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
Django==5.1.1
google-ai-generativelanguage==0.6.9
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.4
json2html==1.3.0
langcodes==3.4.0
language_data==1.2.0
load-dotenv==0.1.0
mailersend==0.5.8
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pillow==10.4.0
preshed==3.0.9
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.8.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sqlparse==0.5.1
srsly==2.4.8
thinc==8.2.5
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
